BACK_END/detector/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
markdown
django-filter
django-cors-headers
torch
einops
timm
opencv-python
numpy
torchvision
Pillow
pandas
facenet-pytorch